Separate Canceled Move Summary into its own component.
