+++ src/scenes/Landing/MoveSummary/CanceledMoveSummary.jsx
@@ -1,0 +1,42 @@
+import React from 'react';
+import { get } from 'lodash';
+import truck from 'shared/icon/truck-gray.svg';
+
+const CanceledMoveSummary = props => {
+  const { profile, reviewProfile } = props;
+  const currentStation = get(profile, 'current_station');
+  const stationPhone = get(currentStation, 'transportation_office.phone_lines.0');
+  return (
+    <div>
+      <h2>New move</h2>
+      <br />
+      <div>
+        <div className="shipment_box">
+          <div className="shipment_type">
+            <img className="move_sm" src={truck} alt="ppm-car" />
+            Start here
+          </div>
+
+          <div className="shipment_box_contents">
+            <div className="step-contents">
+              <div className="status_box usa-width-two-thirds">
+                <div className="step">
+                  <div>
+                    Make sure you have a copy of your move orders before you get started. Questions or need to help?
+                    Contact your local Transportation Office (PPPO) at {get(currentStation, 'name')}
+                    {stationPhone ? ` at ${stationPhone}` : ''}.
+                  </div>
+                </div>
+              </div>
+            </div>
+            <div className="step-links">
+              <button onClick={reviewProfile}>Start</button>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default CanceledMoveSummary;
